Share browser pill editor geometry
