+++ src/app_core/native_shell/composition/state/hit_testing/mod.rs
@@ -8,7 +8,9 @@ mod hover;
 mod map;
 mod waveform;
 
-pub(in crate::app_core::native_shell::composition::state) use self::browser::browser_action_hit_test_cache_key;
+pub(in crate::app_core::native_shell::composition::state) use self::browser::{
+    browser_action_hit_test_cache_key, browser_action_model_signature,
+};
 pub(in crate::app_core::native_shell::composition::state) use self::chrome::sidebar_filter_dropdown_spec;
 pub(in crate::app_core::native_shell::composition::state) use self::map::{
     map_content_id_at_point, map_point_color, map_point_is_focused, map_point_is_selected,
